Use underscored `_index.md` name for index pages (#18)
Update the index page path to be `"_index"`. This better distinguishes the page from the other ordinary pages, and puts the index file naturally at the top of the directory listing, which makes it easier to locate.

Access this value via the `INDEX_PAGE_PATH` constant (renamed from `INDEX_NAME`) and use this constant everywhere we can.