Add basic Tailwind v4 implementation (#55)
Adds a basic implementation for Tailwind CSS v4.x.

The setup follows https://github.com/decafsucks/decafsucks/pull/24 in
that it uses `@tailwindcss/cli` to do the compilation, and then
references the built CSS file in our asset pipeline. This is maybe
_slightly_ unideal – we’d rather run everything through the
`hanami/assets` esbuild pipeline, but when using the
`@tailwindcss/postcss` plugin Tailwind doesn’t detect changes to
template files (meaning we have to restart the server to see change
recognised).

The implementation is super bare-bones for now: it points at the
`templates` directory and it does no customisation to the Tailwind
config.