Add base implementation of a Table of Contents (TOC) scroll indicator (#69)
Addresses #65.

Add a view function and some HTML structure for showing a scroll
indicator alongside the Table of Contents (TOC) on Guide et al pages.
Looks like this:


https://github.com/user-attachments/assets/65e62018-3e78-44a7-9441-915be8f48c61

There’s a few bits that I think we’ll want to do as the design fleshes
out, but this gets the basics in place:

- [ ] Use the finalised flower SVGs (and the correct version per-project
guide)
- [ ] Adjust the position of the flower for different depths. We might
want to actually adjust the SVG slightly so they puff out from the
centre or something?
- [ ] Better design for the indented guides (the additional line feels
awkward)