Add breakpoints utils (#70)
Main addition here is to add a higher-order `breakpointFilter` function
that can be used to wrap any [Defo](https://github.com/makenosound/defo)
view function.

It works like this on the JS side:

```ts
import { breakpointFilter } from "~/utils/breakpoints";
import { tocScrollViewFn } from "./toc-scroll";

export const views: Views = {
  tocScroll: breakpointFilter(tocScrollViewFn),
};
```

Then this on the HTML side:

```erb
<div data-defo-toc-scroll="<%= {
    anchorContainerSelector: ".content",
    breakpoints: ['lg']
  }.to_json %>">
  ...other content
</div>
```

The `breakpoints` array passed will be intercepted by `breakpointFilter`
and the wrapped viewFn will only be called if that breakpoint matches.
If there isn’t a match, it either won’t be called or the existing
instance will be torn down using Defo’s standard `destroy` method.

### Why

For the Table of Contents (TOC) scroll, we want to have pretty different
implementations of that functionality depending on the size of the
user’s device. In fact for the TOC scroll we don’t want it to work at
all on mobile-sized devices (we want to do something completely
different).

### Things to note

- The implementation is tied to [Tailwind’s breakpoint
convention](https://tailwindcss.com/docs/responsive-design), but can
easily be adjusted to fit whatever convention we end up with.
- Tangential commit to ensure the `onScroll` function doesn’t over fire
by debouncing it.