Preload pages with foresight.js (#77)
This PR adds an implementation using
[foresight.js](https://foresightjs.com) to preload links throughout the
docs based on foresight’s heuristics for user-intent. Resolves #72.

### Implementation details

The implementation is pretty straightforward at the moment:

- We initialise the `foresight` view
- It finds all the elements on the page that match our
`elementSelector`, which in this case is `a[^href="/"]` (aka any link
that points to the current root of the current domain).
- We add register those elements with `foresight`

When foresight decides the user is intending to interact with one of
those elements, we do one of two things:

- Issue a `fetch` to get the resolved `href` value
- Check if we’ve already fetched that href and do nothing

Our implementation isn’t responsible for any of the caching itself: that
all defers to the browser’s internal caching mechanisms. All we’re doing
is making that request ahead of time so the browser has it in its cache
before the user actually clicks.

To get benefit from this, we need to ensure the pages we’re requesting
are cached. To do that I’ve added some cache header rules to
`public/_headers`.

### Bundle splitting

Foresight isn’t tiny, and its functionality is a nice-to-have, so I want
to avoid putting it on the hot path. To support this I’ve added the
ability to defer the loading of a module (using dynamic imports) and
split that out in our asset build.

There are two parts to that:

- Adjusting the `esbuild` config to allow for `splitting: true` (we also
need to change the format to `esm` to support this)
- Adding the `lazyLoadView` function that unwraps the import and calls
the view function appropriately