Make images from content/ available in static build (#82)
These were working in local development as of #80, but were not working
as part of the static build generated for deployment.

This is because Parklife's crawling handles web pages only, not any of
the images, etc. that they include.

To address this, add a `content:copy_files` rake task and call it from
`bin/static-build`. This rake task uses the same file and URL patterns
defined in `ContentFileMiddleware` to find image files and copy them to
the appropriate matching locations within `build/`, thereby making those
images available when the site is deployed.

Along with this, update all of our image references to use standard
markdown syntax and `//file` links to ensure a correct absolute path.